webhook/requirements.txt
boto3>=1.34.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
